[edit]Add DateValidationProps and fix accessing LocalizationProvider
