+++ pickers/src/main/kotlin/materialui/pickers/components/DateValidationProps.kt
@@ -1,0 +1,10 @@
+package materialui.pickers.components
+
+import react.RProps
+
+external interface DateValidationProps : RProps {
+    var invalidDateMessage: dynamic
+    var minDateMessage: dynamic
+    var maxDateMessage: dynamic
+    var strictCompareDate: Boolean?
+}
