+++ pickers/src/main/kotlin/materialui/pickers/components/enums/PickerOrientation.kt
@@ -1,0 +1,6 @@
+package materialui.pickers.components.enums
+
+@Suppress("EnumEntryName")
+enum class PickerOrientation {
+    portrait, landscape
+}
